backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pymupdf>=1.24.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
google-generativeai>=0.5.0
pydantic>=2.7.0
